requirements.txt
streamlit
anthropic
openpyxl
pillow
python-dotenv
json-repair